move the build file
