+++ core/exports/mod.ts
@@ -1,4 +1,4 @@
-export { build, mockGlobals } from "../src/build.ts";
+export { build, mockGlobals } from "../src/generate/build.ts";
 export { generateImportMap } from "../src/generate/impormap.ts";
 export { generateManifest, type Manifest } from "../src/generate/manifest.ts";
 export { start } from "../src/start.ts";
